Create Guess
In this game, it will generate a random number and the user has to guess that number. Based on the number the program will tell user if the given number is too high or too low as compared to the random number. At the end, a message is displayed if the user guesses the right number.