Update and rename Guess to Guess the Number
In this app, the User has to Guess a Number which is known only to the programmer. Based on the number, the User will get output such as "High" or "Low" and as per the output the User can trace what number it is and once the correct number is entered the application  will display the output as "Congratulations!!!  This is the Number!!!".