Update Guess the Number
In this, program the programmer sets a number which is only known to him/her. When the user tries to guess the number by entering a random number. The program will tell, the User if the number is "High" or "Low" on the basis of Input given by the programmer. Once the User enters the Correct number, a "message"  will be displayed to the user.