Update Tic-Tac-Toe
In this application, the user can play a Tic-Tac-Toe game in which there will be two players 'X' And 'O'. The players will be able to see the initial board. Based on their turns, they will enter there inputs and once the whole board is filled the app will display "which player has won"  OR  if there is a Tie (Draw).